2019/06/30->pengchenghu：第二阶段验收结束

+++ app/src/main/java/com/automation/zzx/intelligent_basket_demo/utils/http/HttpUtil.java
@@ -176,7 +176,7 @@ public class HttpUtil {
         String command = "/server.command?command=start_rtmp_stream&pipe=0&url=".concat(videoUrl);
         JSONObject jsonObject = new JSONObject();
         jsonObject.put("type", "getVideo");
-        jsonObject.put("device_id", Integer.valueOf(deviceId));
+        jsonObject.put("device_id", Long.valueOf(deviceId));
         jsonObject.put("http_str", command);
         String json = jsonObject.toJSONString();
